+++ NEWS.md
@@ -2,8 +2,8 @@
 News
 ====
 
-Version 0.11.2 - dev
---------------------
+Version 0.12 - dev
+------------------
 
 - NEW: `Insert.block()` returns associated `BlockLayout()` or `None` if block not exist or is an XREF
 - NEW: `Insert.has_scaling` returns `True` if any axis scaling is applied
@@ -13,12 +13,15 @@ Version 0.11.2 - dev
 - NEW: `Insert.explode()` explode block reference entities into target layout (experimental)
 - NEW: `LWPolyline.virtual_entities()` yields 'virtual' LINE and ARC entities
 - NEW: `LWPolyline.explode()` explode LWPOLYLINE as LINE and ARC entities into target layout
+- NEW: `Polyline.virtual_entities()` yields 'virtual' LINE, ARC or 3DFACE entities
+- NEW: `Polyline.explode()` explode POLYLINE as LINE, ARC or 3DFACE entities into target layout
 - NEW: `Block()` properties `is_anonymous`, `is_xref` and `is_xref_overlay`
 - NEW: `R12FastStreamWriter.add_polyline_2d()`, add 2D POLYLINE with start width, end width and bulge value support
 - NEW: `Ellipse.minor_axis` property returns minor axis as `Vector`
 - CHANGE: `R12FastStreamWriter.add_polyline()`, add 3D POLYLINE only, closed flag support
 - CHANGE: renamed `Insert.ucs()` to `Insert.brcs()` which now returns a `BRCS()` object
 - BUGFIX: floating point precision error in `intersection_line_line_2d()`
+- BUGFIX: attribute error in `Polyline.transform_to_wcs()` for 2d polylines
 
 Version 0.11.1 - 2020-02-29
 ---------------------------